backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv==1.0.0
elasticsearch==8.11.0
google-generativeai==0.8.5
google-cloud-firestore==2.21.0
google-auth>=2.40.0
google-api-core>=2.25.1
pydantic>=2.5.0
python-multipart==0.0.6
httpx>=0.28.1
anyio>=4.8.0

